This commit was generated by cvs2svn to compensate for changes in r12175, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@12176 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/maincpu.c
@@ -401,8 +401,8 @@ int maincpu_snapshot_write_module(snapshot_t *s)
         || SMW_B(m, MOS6510_REGS_GET_X(&maincpu_regs)) < 0
         || SMW_B(m, MOS6510_REGS_GET_Y(&maincpu_regs)) < 0
         || SMW_B(m, MOS6510_REGS_GET_SP(&maincpu_regs)) < 0
-        || SMW_W(m, MOS6510_REGS_GET_PC(&maincpu_regs)) < 0
-        || SMW_B(m, MOS6510_REGS_GET_STATUS(&maincpu_regs)) < 0
+        || SMW_W(m, (WORD)MOS6510_REGS_GET_PC(&maincpu_regs)) < 0
+        || SMW_B(m, (BYTE)MOS6510_REGS_GET_STATUS(&maincpu_regs)) < 0
         || SMW_DW(m, (DWORD)last_opcode_info) < 0)
         goto fail;
